updated exporter.R for ve.export

+++ sources/tools/exporter.R
@@ -1,68 +1,89 @@
-# Load the visioneval library to read data
+# Use namespaces to bring in library functions
 requireNamespace("visioneval")
 requireNamespace("tcltk")
 requireNamespace("data.table")
 
-DatastoreType="RD"
-ve.export <- function (outputFolder="output",
-                       includeTables=character(0), # default: all of them
-                       excludeTables=c("Vehicle") # Vehicle doesn't work for VERPAT
-                       ) {
-  #For RPAT includeTables=c("Azone", "Bzone", "Marea", "FuelType", "IncomeGroup","Business","Household")
+# To use this in a visioneval runtime, just do this:
+#   import::here(ve.export,"tools/exporter.R")
 
+# ve.export
+# Dump a model run Datastore to .csv files in an output folder
+#
+# Note that it's not a COMPLETE dump - you don't get base years for
+# tables that were provided as model run inputs (you do get the
+# future year outputs). Some tables get gratuitous name changes.
+# Some tables (e.g. 'Vehicle' in VERPAT) embrace columns of
+# incompatible lengths.
+#
+# Parameters:
+#   modelStateFile="ModelState.Rda" :
+#     Path to ModelState.rda, or of a folder containing ModelState.rda,
+#     or to another file in a folder containing ModelState.rda
+#   outputFolder="output"
+#     Path to folder in which to put the .csv files
+#     NOTE: will be deleted if overwrite=TRUE
+#   includeTables=character(0)
+#     Character vector of table names to include from the Datastore
+#     character(0) or "" will include all
+#   excludeTables=c("Vehicle")
+#     Character vector of table names to exclude from the Datastore
+#     'Model' will always be excluded
+#     'Vehicle' from VERPAT includes incompatible numbers of rows
+#   overwrite=TRUE
+#     Blow away the outputFolder if it already exists
+#   quiet=FALSE
+#     If FALSE, dish up some progress messages and extra debugging data
+#     If TRUE, only report errors
+
+ve.export <- function (modelStateFile = "ModelState.Rda",
+                       outputFolder   = "output",
+                       includeTables  = character(0),  # default: all of them
+                       excludeTables  = c("Vehicle"),  # Vehicle doesn't work for VERPAT
+                       overwrite      = FALSE,         # TRUE to destroy outputFolder first if it exists
+                       quiet          = FALSE          # TRUE to suppress all 
+                       ) {
   owd <- getwd()
   # Do the actual export
-  msf <- getModelStateFile()
+  msf <- getModelStateFile(modelStateFile)
   idx <- indexModelState(msf)
   odt <- outputData(idx)
   exportData(  outputFolder  = outputFolder,
                includeTables = includeTables,
                excludeTables = excludeTables,
+               overwrite     = overwrite,
+               quiet         = quiet,
                odt
              )
   if ( exists("ModelState_ls",envir=.GlobalEnv) ) rm(ModelState_ls,pos=.GlobalEnv) # Hack to support framework
   setwd(owd)
 }
 
-#################################################################
-# Define interior functions
-# (it would be nice just to stuff these into some namespace)
-
-# Helper function to find the directory with ModelState.Rda
-# See: https://stackoverflow.com/questions/48218491/os-independent-way-to-select-directory-interactively-in-r
-choose.directory = function(start.in=getwd(),caption = 'Select model run directory') {
-  tcltk::tk_choose.dir(default=start.in,caption = caption)
-  # The Windows dialog annoyingly won't start in folders that it
-  # unilaterally deems "out of bounds"
-  #   if ( interactive() ) {
-  #     if (.Platform$OS.type == "windows") {
-  #       choose.dir(default=start.in,caption = caption)
-  #     } else {
-  #       tk_choose.dir(default=start.in,caption = caption)
-  #     }
-  #   }
-}
-
 # Locate Model State File
-getModelStateFile <- function() {
-  model.dir <- getwd()
-  modelState <- "ModelState.Rda" # visioneval says you don't get to change this name
-  while ( ! file.exists(modelState) ) {
-    cat("ModelState.Rda not found in",model.dir,"\n")
-    cont <- readline("Pick a different directory? ")
-    if ( substr(cont,1,1) %in% c("N","n","F","f") ) break
-    model.dir <- choose.directory(start.in=model.dir)
-    modelState<-file.path(model.dir,"ModelState.Rda")
-  }
-  if ( ! file.exists(modelState) ) {
-    modelState <- ""
+# See: https://stackoverflow.com/questions/48218491/os-independent-way-to-select-directory-interactively-in-r
+getModelStateFile <- function(modelState) {
+  if ( dir.exists(modelState) ) {
+    model.dir <- modelState
+  } else if ( file.exists(modelState) ) {
+    model.dir <- dirname(modelState)
   } else {
-    modelState <- normalizePath(modelState)
+    model.dir <- "."
   }
+  model.dir <- normalizePath(model.dir)
+  modelState <- file.path(model.dir,"ModelState.Rda") # visioneval says you don't get to change this name
+  if ( interactive() ) {
+    while ( ! file.exists(modelState) ) {
+      cat("ModelState.Rda not found in",model.dir,"\n")
+      cont <- readline("Pick a different directory? ")
+      if ( substr(cont,1,1) %in% c("N","n","F","f") ) break
+      model.dir <- tcltk::tk_choose.dir(default=model.dir,caption = "Select model run directory" )
+      modelState<-file.path(model.dir,"ModelState.Rda")
+    }
+  }
+  if ( ! file.exists(modelState) ) modelState <- ""
 
   # Status messaging and context check
-  cat("Model State File:",modelState,"\n")
-  if ( nchar(modelState)==0 ) stop("No Model State selected.")
+  if ( nchar(modelState)==0 ) stop("No Model State found.")
+  if ( ! quiet ) cat("Model State File:",modelState,"\n")
 
   # Move to location of Model State file (the model run root directory)
   output.dir <- dirname(modelState)
@@ -81,37 +102,6 @@ attributeExist <- function(variable, attr_name){
     }
   }
   FALSE
-}
-
-# NOTE: The following function creates a (standard) data.frame
-makeDataFrame <- function(Table, Output){
-  OutputAllYr <- data.frame()
-
-  for ( year in getYears()){
-    # Looping over each year's worth of data in the given table
-    # OutputIndex is a vector of booleans true for each row in GroupTableName that
-    # contains data for the sought table in the sought year
-    OutputIndex <- Output$GroupTableName$Table %in% Table & Output$GroupTableName$Group %in% year
-    OutputRows  <- Output$Data[OutputIndex] # trim OutputData down to the elements in Table/year
-
-    if ( Table %in% c('Azone', 'Bzone', 'Marea') ){
-      orn <- paste0(Output$GroupTableName$Name[OutputIndex], "_",
-                              Output$Attr[OutputIndex], "_")
-    } else if ( Table %in% c('FuelType', 'IncomeGroup') ){
-      orn <- paste0(Output$GroupTableName$Name[OutputIndex])
-    } else {
-      orn <- paste0(Output$GroupTableName$Name[OutputIndex])
-    }
-    names(OutputRows) <- orn
-    OutputRows <- data.frame(OutputRows, stringsAsFactors = FALSE)
-
-    # Do not output base year except for the Azone, Bzone and Marea Tables
-    if( Table %in% c('Azone', 'Bzone', 'Marea') || year != Output$BaseYear ){
-      OutputRows$Year <- year
-      OutputAllYr <- data.table::rbindlist(list(OutputAllYr, OutputRows), fill = TRUE)
-    }
-  }
-  OutputAllYr
 }
 
 indexModelState <- function(modelState) {
@@ -160,21 +150,60 @@ outputData <- function(Output) {
   }
 
   # Extract the data
-  cat("Extracting data...")
+  if ( ! quiet ) cat("Extracting data...")
   Data<-apply(Output$GroupTableName, 1, readFromTable)
   Attr<-lapply(Data, function(x) attr(x, "UNITS"))
-  cat("Done\n")
+  if ( ! quiet ) cat("Done\n")
   c(Output,list(Data=Data,Attr=Attr))
+}
+
+# NOTE: The following function creates a (standard) data.frame
+makeDataFrame <- function(Table, Output){
+  OutputAllYr <- data.frame()
+
+  for ( year in getYears()){
+    # Looping over each year's worth of data in the given table
+    # OutputIndex is a vector of booleans true for each row in GroupTableName that
+    # contains data for the sought table in the sought year
+    OutputIndex <- Output$GroupTableName$Table %in% Table & Output$GroupTableName$Group %in% year
+    OutputRows  <- Output$Data[OutputIndex] # trim OutputData down to the elements in Table/year
+
+    if ( Table %in% c('Azone', 'Bzone', 'Marea') ){
+      orn <- paste0(Output$GroupTableName$Name[OutputIndex], "_",
+                              Output$Attr[OutputIndex], "_")
+    } else if ( Table %in% c('FuelType', 'IncomeGroup') ){
+      orn <- paste0(Output$GroupTableName$Name[OutputIndex])
+    } else {
+      orn <- paste0(Output$GroupTableName$Name[OutputIndex])
+    }
+    names(OutputRows) <- orn
+    OutputRows <- data.frame(OutputRows, stringsAsFactors = FALSE)
+
+    # Do not output base year except for the Azone, Bzone and Marea Tables
+    if( Table %in% c('Azone', 'Bzone', 'Marea') || year != Output$BaseYear ){
+      OutputRows$Year <- year
+      OutputAllYr <- data.table::rbindlist(list(OutputAllYr, OutputRows), fill = TRUE)
+    }
+  }
+  OutputAllYr
 }
 
 exportData <- function(Output,
                        outputFolder="output",
                        includeTables=character(0),
-                       excludeTables=character(0) ) {
+                       excludeTables=character(0),
+                       overwrite=FALSE,
+                       quiet=TRUE) {
 
   # (Re-)Create the output directory (relative to getwd())
-  if ( dir.exists(outputFolder) || file.exists(outputFolder) )
-    unlink(outputFolder,recursive=TRUE,force=TRUE)
+  if ( dir.exists(outputFolder) || file.exists(outputFolder) ) {
+    if ( overwrite ) {
+      unlink(outputFolder,recursive=TRUE,force=TRUE)
+    } else {
+      cat("Output folder exists:",outputFolder,"\n")
+      stop("Must specify 'overwrite=TRUE' if output folder already exists.")
+    }
+  }
   dir.create(outputFolder)
 
   # Get the list of tables
@@ -186,22 +215,22 @@ exportData <- function(Output,
   } else {
     Tables <- setdiff(Tables,"Model")
   }
-  cat("Include Tables:\n")
-  print(includeTables)
-  cat("Exclude Tables:\n")
-  print(excludeTables)
-  cat("Exporting Tables:\n")
-  print(Tables)
+  if ( ! quiet ) {
+    cat("Include Tables:\n")
+    print(includeTables)
+    cat("Exclude Tables:\n")
+    print(excludeTables)
+    cat("Exporting Tables:\n")
+    print(Tables)
+  }
 
   for ( tbl in Tables ){
-    cat('Exporting', tbl,'...')
+    if ( ! quiet ) cat('Exporting', tbl,'...')
     OutDf <- makeDataFrame(tbl, Output)
     if ( tbl == "IncomeGroup" ) tbl <- "JobAccessibility"
 
     filename <- file.path(outputFolder, paste0(tbl, ".csv"))
     data.table::fwrite(OutDf, file=filename)
-    cat("Done (as",basename(filename),")\n")
+    if ( ! quiet ) cat("Done (as",basename(filename),")\n")
   }
 }
-# End of interior functions
-#################################################################
